requirements.txt
   numpy==1.21.0
   pandas==1.3.0
   tensorflow==2.5.0
   opencv-python==4.5.1.48
   flask==2.0.1
   jinja2
   werkzeug
   pillow
